Fixed a type error

+++ scripts/track_wrangle_silence.py
@@ -247,9 +247,9 @@ if __name__ == "__main__":
     parser.add_argument("output", help="output csv filename or path")
     parser.add_argument("--root-dir-npz", help="set different directory to save npz files")
     parser.add_argument("--root-dir-mp3", help="set different directory to find mp3 files")
-    parser.add_argument("--filter-trim-length", type=int, default=0, help="keep only tracks whose effective length (in seconds) is longer than the theshold")
-    parser.add_argument("--filter-tot-silence-duration", type=int, default=0, help="keep only tracks whose total silent length is shorter than the theshold")
-    parser.add_argument("--filter-max-silence-duration", type=int, default=0, help="keep only tracks whose maximal silent length is shorter than the theshold")
+    parser.add_argument("--filter-trim-length", type=float, default=0, help="keep only tracks whose effective length (in seconds) is longer than the theshold")
+    parser.add_argument("--filter-tot-silence-duration", type=float, default=0, help="keep only tracks whose total silent length is shorter than the theshold")
+    parser.add_argument("--filter-max-silence-duration", type=float, default=0, help="keep only tracks whose maximal silent length is shorter than the theshold")
 
     args = parser.parse_args()
